+++ src/scag/sessions/base/Session2.h
@@ -369,7 +369,7 @@ private:
     /// --- property Scopes
 
     /// 0 -- not set
-    int32_t nextContextId_;                          // pers
+    int16_t nextContextId_;                          // pers
 
     /// global var Scope
     SessionPropertyScope* globalScope_;              // pers
